// server/vance-addon-brain-links/client/src/generated/links/LinksView.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.links.LinksView
*/

import { LinkEntryView } from './LinkEntryView';

export interface LinksView {
  folder: string;
  title?: string;
  description?: string;
  groups: string[];
  entries: LinkEntryView[];
}